Detect YaCy peer version at the end of init as we may fail and prevent preference keys to be created properly.
detectVersion function uses getBaseURL which uses preference keys that are not been created yet. If the function fails,
the init will end and some preference keys will not be created so let the YaCy version be detected once all the keys are created.

Note that it seems the init code should be reviewed and version should be detected dynamically as it depends on the current defined peer.